Improve UI feedback on playlist modification
When adding single tracks to a playlist from a nav buffer, move the
point to the next line afterwards. This provides some feedback to the
user that the interaction was successful as well as allowing adding of
multiple tracks in rapid succession by repeatedly pressing `a`. It
also prevents accidental duplicate adds. This behavior is in line with
that of e.g. the `dired-mark` command in dired buffers.

When adding multiple tracks via a region to a playlist, deactivate the
mark afterwards to provide some feedback to the user in this case,
too, as well as preventing accidental duplicate adds. This too is in
line with the behavior of modes like dired.