Deactivate mark after deleting playlist items
Without this patch there is wonky behavior when deleting multiple
tracks via a region with `transient-mark-mode` from a playlist. If the
last track was included in the region, the whole buffer would end up
getting marked. Otherwise, the end mark would be removed but the start
mark would be left in place.