display file name if missing song name metadata
HTTP streams defined in playlist files do not contain metadata until
_after_ the stream begins playing.  Therefore, it can be difficult to
distinguish between the streams when nothing is displayed for each entry
in the playlist.  I cannot foresee an instance where the file name is
missing, but this may be failure of my imagination.

Therefore, if the song title is missing, display the file name instead.

This behaviour is consistent with `ncmpcpp` sans special formatting
which `ncmpcpp` does to tracks starting with "http...".

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>